+++ epoxy-integrationtest/src/main/java/com/airbnb/epoxy/AdapterWithIdChanged.java
@@ -1,0 +1,11 @@
+package com.airbnb.epoxy;
+
+class AdapterWithIdChanged extends EpoxyController {
+
+  @AutoModel Model_ model1 = new Model_();
+
+  @Override
+  protected void buildModels() {
+    add(model1.id(23));
+  }
+}
