+++ epoxy-integrationtest/src/main/java/com/airbnb/epoxy/ControllerWithAutoModel.java
