+++ epoxy-integrationtest/src/test/java/com/airbnb/epoxy/AutoModelIntegrationTest.java
@@ -1,7 +1,5 @@
 package com.airbnb.epoxy;
 
-import android.view.View;
-
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.robolectric.RobolectricTestRunner;
@@ -15,26 +13,6 @@ import static org.junit.Assert.assertEquals;
 @Config(constants = BuildConfig.class, sdk = 21)
 public class AutoModelIntegrationTest {
 
-  static class TestModel extends EpoxyModel<View> {
-
-    @Override
-    protected int getDefaultLayout() {
-      return 0;
-    }
-  }
-
-  static class BasicAutoModelsAdapter extends EpoxyController {
-
-    @AutoModel TestModel model1;
-    @AutoModel TestModel model2;
-
-    @Override
-    protected void buildModels() {
-      add(model1.id(1));
-      add(model2.id(2));
-    }
-  }
-
   @Test
   public void basicAutoModels() {
     BasicAutoModelsAdapter controller = new BasicAutoModelsAdapter();
@@ -43,34 +21,14 @@ public class AutoModelIntegrationTest {
     List<EpoxyModel<?>> models = controller.getAdapter().getCopyOfModels();
 
     assertEquals("Models size", 2, models.size());
-    assertEquals("First model", TestModel.class, models.get(0).getClass());
-    assertEquals("Second model", TestModel.class, models.get(1).getClass());
-  }
-
-  static class AdapterWithFieldAssigned extends EpoxyController {
-
-    @AutoModel TestModel model1 = new TestModel();
-
-    @Override
-    protected void buildModels() {
-      add(model1);
-    }
+    assertEquals("First model", Model_.class, models.get(0).getClass());
+    assertEquals("Second model", Model_.class, models.get(1).getClass());
   }
 
   @Test(expected = IllegalStateException.class)
   public void assigningValueToFieldFails() {
     AdapterWithFieldAssigned testAdapter = new AdapterWithFieldAssigned();
     testAdapter.requestModelBuild();
-  }
-
-  static class AdapterWithIdChanged extends EpoxyController {
-
-    @AutoModel TestModel model1 = new TestModel();
-
-    @Override
-    protected void buildModels() {
-      add(model1.id(23));
-    }
   }
 
   @Test(expected = IllegalStateException.class)
